+++ src/main/java/nordmods/uselessreptile/client/renderer/base/DragonEquipmentRenderer.java
@@ -18,8 +18,7 @@ import software.bernie.geckolib.cache.object.BakedGeoModel;
 import software.bernie.geckolib.renderer.GeoObjectRenderer;
 
 public class DragonEquipmentRenderer extends GeoObjectRenderer<DragonEquipmentAnimatable> {
-    private float yaw = 0;
-    private float roll = 0;
+
     public DragonEquipmentRenderer() {
         super(new DragonEqupmentModel());
         addRenderLayer(new URGlowingLayer<>(this));
@@ -34,17 +33,18 @@ public class DragonEquipmentRenderer extends GeoObjectRenderer<DragonEquipmentAn
 
         URDragonEntity owner = animatable.owner;
 
-        float yaw = MathHelper.lerpAngleDegrees(partialTick, owner.prevBodyYaw, owner.bodyYaw);
-        if (owner.isFrozen())
-            yaw += (float)(Math.cos(owner.age * 3.25d) * Math.PI * 0.4d);
-        this.yaw = 180f - yaw;
-        poseStack.multiply(RotationAxis.POSITIVE_Y.rotationDegrees(this.yaw));
+        if (!isReRender) {
+            float yaw = MathHelper.lerpAngleDegrees(partialTick, owner.prevBodyYaw, owner.bodyYaw);
+            if (owner.isFrozen())
+                yaw += (float) (Math.cos(owner.age * 3.25d) * Math.PI * 0.4d);
+            yaw = 180f - yaw;
+            poseStack.multiply(RotationAxis.POSITIVE_Y.rotationDegrees(yaw));
 
-        this.roll = 0;
-        if (owner.deathTime > 0) {
-            float deathRotation = (owner.deathTime + partialTick - 1f) / 20f * 1.6f;
-            this.roll = Math.min(MathHelper.sqrt(deathRotation), 1) * 90f;
-            poseStack.multiply(RotationAxis.POSITIVE_Z.rotationDegrees(this.roll));
+            if (owner.deathTime > 0) {
+                float deathRotation = (owner.deathTime + partialTick - 1f) / 20f * 1.6f;
+                float roll = Math.min(MathHelper.sqrt(deathRotation), 1) * 90f;
+                poseStack.multiply(RotationAxis.POSITIVE_Z.rotationDegrees(roll));
+            }
         }
 
         scaleWidth = scaleHeight = owner.getScale();
@@ -52,14 +52,11 @@ public class DragonEquipmentRenderer extends GeoObjectRenderer<DragonEquipmentAn
         poseStack.translate(0, 0.01, 0);
     }
 
-    //todo fix z-fighting
     //*screams*
     @Override
     public void preApplyRenderLayers(MatrixStack poseStack, DragonEquipmentAnimatable animatable, BakedGeoModel model, RenderLayer renderType, VertexConsumerProvider bufferSource,
                                      VertexConsumer buffer, float partialTick, int packedLight, int packedOverlay) {
         poseStack.push();
-        poseStack.multiply(RotationAxis.NEGATIVE_Z.rotationDegrees(roll));
-        poseStack.multiply(RotationAxis.NEGATIVE_Y.rotationDegrees(yaw));
         poseStack.translate(0, -0.01, 0);
         super.preApplyRenderLayers(poseStack, animatable, model, renderType, bufferSource, buffer, partialTick, packedLight, packedOverlay);
         poseStack.pop();
@@ -69,8 +66,6 @@ public class DragonEquipmentRenderer extends GeoObjectRenderer<DragonEquipmentAn
     public void applyRenderLayers(MatrixStack poseStack, DragonEquipmentAnimatable animatable, BakedGeoModel model, RenderLayer renderType, VertexConsumerProvider bufferSource,
                                   VertexConsumer buffer, float partialTick, int packedLight, int packedOverlay) {
         poseStack.push();
-        poseStack.multiply(RotationAxis.NEGATIVE_Z.rotationDegrees(roll));
-        poseStack.multiply(RotationAxis.NEGATIVE_Y.rotationDegrees(yaw));
         poseStack.translate(0, -0.01, 0);
         super.applyRenderLayers(poseStack, animatable, model, renderType, bufferSource, buffer, partialTick, packedLight, packedOverlay);
         poseStack.pop();